requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
pyTelegramBotAPI
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
dateparser
